cleanup StatusCell code
